# swagger.yaml
swagger: "2.0"
host: localhost:8080
basePath: /
produces:
    - application/json
consumes:
    - application/json
schemes:
    - http
info:
    description: NFL application
    title: NFL application
    version: 0.0.0
definitions:
    player:
        type: object
        required:
            - name
        properties:
            id:
                type: integer
                format: int64
                readOnly: true
            name:
                type: string
                minLength: 1
            avatar:
                type: string
            created_at:
                type: string
            updated_at:
                type: string
            deleted_at:
                type: string
    error:
        type: object
        required:
            - message
        properties:
            message:
                type: string
paths:
    /players:
        get:
            tags:
                - players
            operationId: getList
            parameters:
                -   name: page
                    in: query
                    type: integer
                    format: int64
                -   name: per_page
                    in: query
                    type: integer
                    format: int32
                    default: 2
            responses:
                200:
                    description: All players
                    schema:
                        properties:
                            data:
                                $ref: "#/definitions/player"
                        example:
                            data:
                                -   id: 1
                                    name: "EXAMPLE_PLAYER_1"
                                    avatar: "f9fa4c7f-74dd-4a8d-8613-c5017fe047c9.jpg"
                                    created_at: "2018-12-14T11:44:32.779195Z"
                                    updated_at: "2018-12-17T12:59:44.153986Z"
                                    deleted_at: null
                                -   id: 2
                                    name: "EXAMPLE_PLAYER_2"
                                    avatar: "c5017fe047c9-74dd-4a8d-8613-f9fa4c7f47c9.jpg"
                                    created_at: "2018-12-14T11:44:32.779195Z"
                                    updated_at: "2018-12-17T12:59:44.153986Z"
                                    deleted_at: null
                default:
                    description: error
                    schema:
                        $ref: "#/definitions/error"
        post:
            tags:
                - players
            operationId: add
            parameters:
                -   name: name
                    description: "Player name"
                    in: body
                    schema:
                        example:
                            name: "EXAMPLE_PLAYER_1"
            responses:
                201:
                    description: Created
                    schema:
                        properties:
                            data:
                                $ref: "#/definitions/player"
                        example:
                            data:
                                id: 2
                                name: "EXAMPLE_PLAYER_1"
                                avatar: ""
                                created_at: "2018-12-14T11:44:32.779195Z"
                                updated_at: "2018-12-17T12:59:44.153986Z"
                                deleted_at: null
                default:
                    description: error
                    schema:
                        $ref: "#/definitions/error"
    /players/{id}:
        parameters:
            -   type: integer
                format: int64
                name: id
                in: path
                required: true
        get:
            tags:
                - players
            operationId: get
            responses:
                200:
                    description: Get single
                    schema:
                        properties:
                            data:
                                $ref: "#/definitions/player"
                        example:
                            data:
                                id: 2
                                name: "EXAMPLE_PLAYER_1"
                                avatar: ""
                                created_at: "2018-12-14T11:44:32.779195Z"
                                updated_at: "2018-12-17T12:59:44.153986Z"
                                deleted_at: null
                default:
                    description: error
                    schema:
                        $ref: "#/definitions/error"
        delete:
            tags:
                - players
            operationId: delete
            responses:
                204:
                    description: Deleted
                default:
                    description: error
                    schema:
                        $ref: "#/definitions/error"